The Student registration functionality is now complete.
You register a Student by inputing a Student object and a StudentData
object. The StudentData object stores information about the Student. It
is composed of EducationBackground objects, a Biodata object,
ProfessionalExperience objects and MeansOfDiscovery objects.

Internally, the StudentManager contains a private class
StudentDataManager that does the insertion of the StudentData object.

The StudentManager method insert() inserts both the StudentData
attributes and the Student object as one atomic operation