Getting list of tables from database
The functionality of retrieving tables from the database as pagged
output by specifying only the start index is
about to be implemented in each manager class

The idea is to create method getAll(int startIndex) in each the manager
class. This method would execute a query that retrieves the rows of the
table from the database which have a limit of 30 rows. This would ensure
that extra memory is not wasted/demanded