Method getAll has been added for all the manager classes but not tested.
Making a call to this method would return 30 objects as is in the table.
They all accept a startIndex which controls where the retrieval would
begin. This can be used to create a paging application where the
startIndex would be increased by 30 in each iteration. 

I am also thinking of adding method getCount that would return the total
number of items in each table thus making the number of pages easy to
calculate.

Note that in this commit some errors are found in the
ModularRegisterManager class. It was discovered that the update method
of this class has not yet been implemented. This would be worked on in
subsequent commits