Functionality retrieve log has been added but not tested
Logs information can be retrieved from the LogManger class method
 getLog by specifying the subclass of Log to be retrieved, the LogType
and the startIndex
 
 The LogType enum contains constants INSERT, UPDATE and DELETE which
tells the logManager class to retrieve a soecific type of log from the
Log subclass
 
 This method can be used to retrieve any information in the log tables
in the application
 
 
 Also in this commit the getImageFromStream method in class
StudentManager has been slightly modified to ignore emppty files and in
the database. This means that if a Student does not have and image then
a nulll File would be saved as the Student's image attribute. This
feature has also been tested working
 
 The next step is to add some comments to the code and test the
LogManager class