Commenting of Log tables currently on-going.
I found a missing feature that i would like to add into the system.
There is currently no way of knowing what an object was updated from in
the log tables because it stores only what it was updated to.

That functionality is about to be added to the system