Log tables contains more info about transactions
I have added functionality to the log tables that makes it possible to
know what object what deleted, inserted, updated from and updated to.

This functionality has been lightly tested but i would like to resume
commenting of the code before i come back to it again. 

Once i am done with the logging functionality and comments, i would dive
into making the PaymentManager more useful. 

The end of this is near