Log tables are being tested and almost all are okay
The Log tables are currently being tested and most of them are working.
However the StudentLog and PaymentLog have not been confirmed working
because of a bug found in the StudentManager and PaymentManager
These would need to be fixed before we can confirm their individual
logs. 
For now however, commenting of the code continues.