Payment functionality has been added but not tested
It is now possible to make payment for a registered module. 
However, this operation has not been tested.

The next step from here is to conduct a thorough test on all the
operaions that have been added to the system thus far, It would be done
by first emptying all the data in the database