Class StudentModuleStats has been added to retrieve statistic about the individual modules a student has registered for.
Modification to PaymentManager class was also made. It now correctly
performs the calculation of the remaining payment which was wrongly
claculated as a result of how double numbers are stored. 

The required changes were made to the StatsTest class and the new
functionality has been confirmed working. 

More test are required though